Rename omsa to setup-hw
Remove setup-hw command.
