--- pkg/setup-hw/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/cybozu-go/neco/pkg/setup-hw/cmd"
-
-func main() {
-	cmd.Execute()
-}
