+++ progs/setuphw/service.go
@@ -1,4 +1,4 @@
-package omsa
+package setuphw
 
 import (
 	"io"
